feat(260607-k1n): tolerate decimal quantities from LLM receipt parser
- ReceiptItem.Quantity widened int → float64 so JSON 40.0 unmarshals
- createPurchaseEvent INSERT rounds float64 → int via math.Round at the
  DB-write boundary (purchase_line_items.quantity stays INTEGER)
- ValidateReceiptData updated for the new type (drop float64() cast on
  Check 2; round totalQty in Check 3 to match int summary semantics)
- Add TestParseJSONBody_DecimalQuantity regression test exercising
  today's exact Restaurant Depot $391.96 failing payload

inventory.PurchaseLineItemInput/Output Quantity stays int (FE/DB-facing).
No migrations, no FE changes, no new deps.

+++ backend/internal/receipt/types.go
@@ -35,9 +35,15 @@ type Attachment struct {
 }
 
 // ReceiptItem is a single line item parsed from a receipt.
+//
+// Quantity is float64 so JSON values like `40.0` (which Anthropic returns for
+// some receipts — e.g. Restaurant Depot, 260607-k1n) unmarshal cleanly instead
+// of failing the strict int decoder. The DB column purchase_line_items.quantity
+// is INTEGER — the int coercion happens at the single DB-write boundary in
+// worker.go (createPurchaseEvent) via math.Round.
 type ReceiptItem struct {
 	Name     string  `json:"name"`
-	Quantity int     `json:"quantity"`
+	Quantity float64 `json:"quantity"`
 	Price    float64 `json:"price"`
 	IsCase   bool    `json:"is_case"`
 }
